+++ packages/expo-cli/src/commands/upload/IOSUploader.ts
@@ -83,8 +83,8 @@ export type IosPlatformOptions = PlatformOptions & {
   appName: string;
   language?: string;
   appleTeamId?: string;
-  itcTeamId?: string;
   publicUrl?: string;
+  companyName?: string;
 };
 
 type AppleCreds = Pick<IosPlatformOptions, 'appleId' | 'appleIdPassword'>;
@@ -115,7 +115,7 @@ export default class IOSUploader extends BaseUploader {
     const appleIdCrentials = await this._getAppleIdCredentials();
     const appleTeamId = await this._getAppleTeamId();
     const appName = await this._getAppName();
-    const otherOptions = pick(this.options, ['language', 'sku']);
+    const otherOptions = pick(this.options, ['language', 'sku', 'companyName']);
     return {
       ...appleIdCrentials,
       appName,
@@ -179,20 +179,40 @@ export default class IOSUploader extends BaseUploader {
 
   async _uploadToTheStore(platformData: IosPlatformOptions, buildPath: string): Promise<void> {
     const { fastlane } = this;
-    const { appleId, appleIdPassword, appName, language, appleTeamId } = platformData;
-    const { bundleIdentifier } = this._exp && this._exp.ios;
+    const { appleId, appleIdPassword, appName, language, appleTeamId, companyName } = platformData;
 
-    const appleCreds = { appleId, appleIdPassword, appleTeamId, itcTeamId: this.options.itcTeamId };
+    const appleCreds = { appleId, appleIdPassword, appleTeamId, companyName };
 
-    log('Ensuring the app exists on App Store Connect, this may take a while...');
-    await runFastlaneAsync(
-      fastlane.appProduce,
-      [bundleIdentifier, appName, appleId, language],
+    log('Resolving the ITC team ID...');
+    const { itc_team_id: itcTeamId } = await runFastlaneAsync(
+      fastlane.resolveItcTeamId,
+      [],
       appleCreds
     );
+    log(`ITC team ID is ${itcTeamId}`);
+    const updatedAppleCreds = {
+      ...appleCreds,
+      itcTeamId,
+    };
+
+    log('Ensuring the app exists on App Store Connect, this may take a while...');
+    try {
+      await runFastlaneAsync(
+        fastlane.appProduce,
+        [this._exp?.ios?.bundleIdentifier, appName, appleId, language],
+        updatedAppleCreds
+      );
+    } catch (err) {
+      if (err.message.match(/You must provide a company name to use on the App Store/)) {
+        log.error(
+          'You haven\'t uploaded any app to App Store yet. Please provide your company name with --company-name "COMPANY NAME"'
+        );
+      }
+      throw err;
+    }
 
     log('Uploading the app to Testflight, hold tight...');
-    await runFastlaneAsync(fastlane.pilotUpload, [buildPath, appleId], appleCreds);
+    await runFastlaneAsync(fastlane.pilotUpload, [buildPath, appleId], updatedAppleCreds);
 
     log(
       `All done! You may want to go to App Store Connect (${chalk.underline(