+++ packages/webpack-config/src/env/__tests__/getMode-test.js
@@ -1,3 +1,5 @@
+/* eslint-env node */
+
 import getMode from '../getMode';
 
 let mode;